requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
argon2-cffi==23.1.0
python-dotenv==1.0.0
marshmallow==3.20.1
flask-cors==4.0.0
redis==5.0.1
email-validator==2.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.3.0
